feat: Resizable partly
pref: Rename WIN_WIDTH/HEIGHT to DEF_WIN_WIDTH/HEIGHT

+++ src/ui/utils/rounded.cpp
@@ -72,7 +72,7 @@ namespace sf {
         if (index >= myCornerPointCount * 4)
             return {0, 0};
 
-        float deltaAngle = 90.0f / (myCornerPointCount - 1);
+        float deltaAngle = 90.0f / static_cast<float>(myCornerPointCount - 1);
         sf::Vector2f center;
         unsigned int centerIndex = index / myCornerPointCount;
         static const float pi = 3.141592654f;
@@ -98,7 +98,7 @@ namespace sf {
                 break;
         }
 
-        return {static_cast<float>(myRadius * cos(deltaAngle * (index - centerIndex) * pi / 180) + center.x),
-                static_cast<float>(-myRadius * sin(deltaAngle * (index - centerIndex) * pi / 180) + center.y)};
+        return {static_cast<float>(myRadius * cos(deltaAngle * static_cast<float>(index - centerIndex) * pi / 180) + center.x),
+                static_cast<float>(-myRadius * sin(deltaAngle * static_cast<float>(index - centerIndex) * pi / 180) + center.y)};
     }
 } // namespace sf
